feat: replace sequence NRS with multimap NRS (register)
