+++ Cargo.toml
@@ -32,7 +32,8 @@ rmp-serde = "~0.15"
 serde = "1.0.123"
 serde_json = "1.0.62"
 sha3 = "~0.9"
-safe_network = "0.10.0"
+# safe_network = "0.10.0"
+safe_network = {path = "../safe_network"}
 thiserror = "1.0.23"
 uhttp_uri = "~0.5"
 url = "2.2.0"
