+++ src/app/mod.rs
@@ -30,7 +30,6 @@ pub mod fetch;
 pub mod files;
 pub mod multimap;
 pub mod nrs;
-pub mod nrs_multimap;
 pub mod register;
 pub use consts::DEFAULT_XORURL_BASE;
 pub use helpers::parse_tokens_amount;
