+++ src/app/nrs_seq/mod.rs
@@ -16,11 +16,10 @@ use crate::{
         consts::{CONTENT_ADDED_SIGN, CONTENT_DELETED_SIGN},
         Safe,
     },
-    register::EntryHash,
-    Error, Result, SafeUrl, XorUrl,
+    Error, Result, SafeContentType, SafeUrl, XorUrl,
 };
 use log::{debug, info, warn};
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
 
 // Type tag to use for the NrsMapContainer stored on Register
 pub(crate) const NRS_MAP_TYPE_TAG: u64 = 1_500;
@@ -31,7 +30,7 @@ const ERROR_MSG_NO_NRS_MAP_FOUND: &str = "No NRS Map found at this address";
 pub type ProcessedEntries = BTreeMap<String, (String, String)>;
 
 impl Safe {
-    pub fn mm_parse_url(url: &str) -> Result<SafeUrl> {
+    pub fn parse_url(url: &str) -> Result<SafeUrl> {
         let safe_url = SafeUrl::from_url(&sanitised_url(url))?;
         Ok(safe_url)
     }
@@ -40,17 +39,16 @@ impl Safe {
     // It also returns a second SafeUrl if the URL was resolved from an NRS-URL,
     // this second SafeUrl instance contains the information of the parsed NRS-URL.
     // *Note* this is not part of the public API, but an internal helper function used by API impl.
-    pub(crate) async fn mm_parse_and_resolve_url(
+    pub(crate) async fn parse_and_resolve_url(
         &self,
         url: &str,
     ) -> Result<(SafeUrl, Option<SafeUrl>)> {
-        let safe_url = Safe::mm_parse_url(url)?;
+        let safe_url = Safe::parse_url(url)?;
         let orig_path = safe_url.path_decoded()?;
 
         // Obtain the resolution chain without resolving the URL's path
         let mut resolution_chain = self
             .retrieve_from_url(
-                // TODO take a look at safe url code where its used, ask gab
                 &safe_url.to_string(),
                 false,
                 None,
@@ -77,7 +75,7 @@ impl Safe {
         }
     }
 
-    pub async fn mm_nrs_map_container_add(
+    pub async fn nrs_map_container_add(
         &self,
         name: &str,
         link: &str,
@@ -89,7 +87,7 @@ impl Safe {
         // GET current NRS map from name's TLD
         let (safe_url, _) = validate_nrs_name(name)?;
         let xorurl = safe_url.to_string();
-        let (version, mut nrs_map) = self.mm_nrs_map_container_get(&xorurl).await?;
+        let (version, mut nrs_map) = self.nrs_map_container_get(&xorurl).await?;
         debug!("NRS, Existing data: {:?}", nrs_map);
 
         let link = nrs_map.update(name, link, default, hard_link)?;
@@ -98,29 +96,16 @@ impl Safe {
 
         debug!("The new NRS Map: {:?}", nrs_map);
         if !dry_run {
-            let nrs_map_xorurl = self.mm_store_nrs_map(&nrs_map).await?;
-            let old_values: BTreeSet<EntryHash> = self
-                .fetch_multimap_values(&safe_url)
-                .await?
-                .iter()
-                .map(|(hash, _)| hash.to_owned())
-                .collect();
-            let entry = (
-                name.as_bytes().to_owned(),
-                nrs_map_xorurl.as_bytes().to_owned(),
-            );
-            self.multimap_insert(&xorurl, entry, old_values).await?;
-
-            // TODO rm old impl below kept as a reference
             // Append new version of the NrsMap in the Public Sequence (NRS Map Container)
-            // self.safe_client
-            //     .append_to_sequence(
-            //         nrs_map_xorurl.as_bytes(),
-            //         safe_url.xorname(),
-            //         safe_url.type_tag(),
-            //         false,
-            //     )
-            //     .await?;
+            let nrs_map_xorurl = self.store_nrs_map(&nrs_map).await?;
+            self.safe_client
+                .append_to_sequence(
+                    nrs_map_xorurl.as_bytes(),
+                    safe_url.xorname(),
+                    safe_url.type_tag(),
+                    false,
+                )
+                .await?;
         }
 
         Ok((version + 1, xorurl, processed_entries, nrs_map))
@@ -130,20 +115,21 @@ impl Safe {
     ///
     /// ## Example
     ///
-    /// ```rust_no_run
+    /// ```rust,no_run
     /// # use rand::distributions::Alphanumeric;
     /// # use rand::{thread_rng, Rng};
     /// # use sn_api::Safe;
     /// # let mut safe = Safe::default();
-    /// # async_std::task::block_on(async {
-    /// #   safe.connect("", Some("fake-credentials")).await.unwrap();
+    /// # let rt = tokio::runtime::Runtime::new().unwrap();
+    /// # rt.block_on(async {
+    /// #   safe.connect(None, None, None).await.unwrap();
     ///     let rand_string: String = thread_rng().sample_iter(&Alphanumeric).take(15).collect();
     ///     let file_xorurl = safe.files_store_public_blob(&vec![], None, false).await.unwrap();
-    ///     let (xorurl, _processed_entries, nrs_map_container) = safe.mm_nrs_map_container_create(&rand_string, &file_xorurl, true, false, false).await.unwrap();
+    ///     let (xorurl, _processed_entries, nrs_map_container) = safe.nrs_map_container_create(&rand_string, &file_xorurl, true, false, false).await.unwrap();
     ///     assert!(xorurl.contains("safe://"))
     /// # });
     /// ```
-    pub async fn mm_nrs_map_container_create(
+    pub async fn nrs_map_container_create(
         &mut self,
         name: &str,
         link: &str,
@@ -153,14 +139,14 @@ impl Safe {
     ) -> Result<(XorUrl, ProcessedEntries, NrsMap)> {
         info!("Creating an NRS map");
         let (_, nrs_url) = validate_nrs_name(name)?;
-        if self.mm_nrs_map_container_get(&nrs_url).await.is_ok() {
+        if self.nrs_map_container_get(&nrs_url).await.is_ok() {
             Err(Error::ContentError(
                 "NRS name already exists. Please use 'nrs add' command to add sub names to it"
                     .to_string(),
             ))
         } else {
             let mut nrs_map = NrsMap::default();
-            let link = nrs_map.update(&name, link, default, hard_link)?;
+            let link = nrs_map.update(name, link, default, hard_link)?;
             let mut processed_entries = ProcessedEntries::new();
             processed_entries.insert(name.to_string(), (CONTENT_ADDED_SIGN.to_string(), link));
 
@@ -172,50 +158,35 @@ impl Safe {
                 debug!("XorName for \"{:?}\" is \"{:?}\"", &nrs_url, &nrs_xorname);
 
                 // Store the serialised NrsMap in a Public Blob
-                let nrs_map_xorurl = self.mm_store_nrs_map(&nrs_map).await?;
+                let nrs_map_xorurl = self.store_nrs_map(&nrs_map).await?;
 
-                // Create a new multimap
-                let xorurl = self
-                    .multimap_create(Some(nrs_xorname), NRS_MAP_TYPE_TAG, false)
+                // Store the NrsMapContainer in a Public Sequence, putting the
+                // serialised NrsMap XOR-URL as the first entry value
+                let xorname = self
+                    .safe_client
+                    .store_sequence(
+                        nrs_map_xorurl.as_bytes(),
+                        Some(nrs_xorname),
+                        NRS_MAP_TYPE_TAG,
+                        None,
+                        false,
+                    )
                     .await?;
 
-                // Store the NRS map in the multimap
-                let entry = (
-                    name.as_bytes().to_owned(),
-                    nrs_map_xorurl.as_bytes().to_owned(),
-                );
-                let _ = self
-                    .multimap_insert(&xorurl, entry, BTreeSet::new())
-                    .await?;
-
-                // TODO rm old impl below kept as a reference
-                // // Store the NrsMapContainer in a Public Sequence, putting the
-                // // serialised NrsMap XOR-URL as the first entry value
-                // let xorname = self
-                //     .safe_client
-                //     .store_sequence(
-                //         nrs_map_xorurl.as_bytes(),
-                //         Some(nrs_xorname),
-                //         NRS_MAP_TYPE_TAG,
-                //         None,
-                //         false,
-                //     )
-                //     .await?;
-                //
-                // let xorurl = SafeUrl::encode_sequence_data(
-                //     xorname,
-                //     NRS_MAP_TYPE_TAG,
-                //     SafeContentType::NrsMapContainer,
-                //     self.xorurl_base,
-                //     false,
-                // )?;
+                let xorurl = SafeUrl::encode_sequence_data(
+                    xorname,
+                    NRS_MAP_TYPE_TAG,
+                    SafeContentType::NrsMapContainer,
+                    self.xorurl_base,
+                    false,
+                )?;
 
                 Ok((xorurl, processed_entries, nrs_map))
             }
         }
     }
 
-    pub async fn mm_nrs_map_container_remove(
+    pub async fn nrs_map_container_remove(
         &self,
         name: &str,
         dry_run: bool,
@@ -224,7 +195,7 @@ impl Safe {
         // GET current NRS map from &name TLD
         let (safe_url, _) = validate_nrs_name(name)?;
         let xorurl = safe_url.to_string();
-        let (version, mut nrs_map) = self.mm_nrs_map_container_get(&xorurl).await?;
+        let (version, mut nrs_map) = self.nrs_map_container_get(&xorurl).await?;
         debug!("NRS, Existing data: {:?}", nrs_map);
 
         let removed_link = nrs_map.nrs_map_remove_subname(name)?;
@@ -234,26 +205,18 @@ impl Safe {
             (CONTENT_DELETED_SIGN.to_string(), removed_link),
         );
 
-        debug!("Removing from multimap");
+        debug!("The new NRS Map: {:?}", nrs_map);
         if !dry_run {
-            let old_values: BTreeSet<EntryHash> = self
-                .fetch_multimap_values(&safe_url)
-                .await?
-                .iter()
-                .map(|(hash, _)| hash.to_owned())
-                .collect();
-            self.multimap_remove(&xorurl, old_values).await?;
-
-            // TODO rm old impl below kept as a reference
-            // let nrs_map_xorurl = self.mm_store_nrs_map(&nrs_map).await?;
-            // self.safe_client
-            //     .append_to_sequence(
-            //         nrs_map_xorurl.as_bytes(),
-            //         safe_url.xorname(),
-            //         safe_url.type_tag(),
-            //         false,
-            //     )
-            //     .await?;
+            // Append new version of the NrsMap in the Public Sequence (NRS Map Container)
+            let nrs_map_xorurl = self.store_nrs_map(&nrs_map).await?;
+            self.safe_client
+                .append_to_sequence(
+                    nrs_map_xorurl.as_bytes(),
+                    safe_url.xorname(),
+                    safe_url.type_tag(),
+                    false,
+                )
+                .await?;
         }
 
         Ok((version + 1, xorurl, processed_entries, nrs_map))
@@ -263,73 +226,58 @@ impl Safe {
     ///
     /// ## Example
     ///
-    /// ```rust_no_run
+    /// ```rust,no_run
     /// # use sn_api::Safe;
     /// # use rand::distributions::Alphanumeric;
     /// # use rand::{thread_rng, Rng};
     /// # let mut safe = Safe::default();
-    /// # async_std::task::block_on(async {
-    /// #   safe.connect("", Some("fake-credentials")).await.unwrap();
+    /// # let rt = tokio::runtime::Runtime::new().unwrap();
+    /// # rt.block_on(async {
+    /// #   safe.connect(None, None, None).await.unwrap();
     ///     let rand_string: String = thread_rng().sample_iter(&Alphanumeric).take(15).collect();
     ///     let file_xorurl = safe.files_store_public_blob(&vec![], Some("text/plain"), false).await.unwrap();
-    ///     let (xorurl, _processed_entries, _nrs_map) = safe.mm_nrs_map_container_create(&rand_string, &file_xorurl, true, false, false).await.unwrap();
-    ///     let (version, nrs_map_container) = safe.mm_nrs_map_container_get(&xorurl).await.unwrap();
+    ///     let (xorurl, _processed_entries, _nrs_map) = safe.nrs_map_container_create(&rand_string, &file_xorurl, true, false, false).await.unwrap();
+    ///     let (version, nrs_map_container) = safe.nrs_map_container_get(&xorurl).await.unwrap();
     ///     assert_eq!(version, 0);
     ///     assert_eq!(nrs_map_container.get_default_link().unwrap(), file_xorurl);
     /// # });
     /// ```
-    pub async fn mm_nrs_map_container_get(&self, url: &str) -> Result<(u64, NrsMap)> {
+    pub async fn nrs_map_container_get(&self, url: &str) -> Result<(u64, NrsMap)> {
         debug!("Getting latest resolvable map container from: {:?}", url);
-        let safe_url = Safe::mm_parse_url(url)?;
+        let safe_url = Safe::parse_url(url)?;
 
-        // TODO: versions should be managed with EntryHash instead of u64
-        // since this requires to upgrade the safe_url API, we kept is as is for now
-        let placeholder_version: u64 = 0;
-        // TODO: manage multiple resolutions currently only returns the 1st one
-        // fetch multimap latest values
-        let data = match self.fetch_multimap_values(&safe_url).await?.iter().next() {
-            Some((_version_hash, (_name, nrs_map_xorurl_bytes))) => {
-                Ok((placeholder_version, nrs_map_xorurl_bytes.to_owned()))
-            }
-            None => Err(Error::EmptyContent(format!(
-                "Empty Sequence found at XoR name {}",
-                safe_url.xorname()
-            ))),
-        };
-
-        // TODO rm old impl below kept as a reference
         // Check if the URL specified a specific version of the content or simply the latest available
-        // let data = match safe_url.content_version() {
-        //     None => {
-        //         self.safe_client
-        //             .sequence_get_last_entry(safe_url.xorname(), NRS_MAP_TYPE_TAG, false)
-        //             .await
-        //     }
-        //     Some(content_version) => {
-        //         let serialised_nrs_map = self
-        //             .safe_client
-        //             .sequence_get_entry(
-        //                 safe_url.xorname(),
-        //                 NRS_MAP_TYPE_TAG,
-        //                 content_version,
-        //                 false,
-        //             )
-        //             .await
-        //             .map_err(|_| {
-        //                 Error::VersionNotFound(format!(
-        //                     "Version '{}' is invalid for NRS Map Container found at \"{}\"",
-        //                     content_version, url,
-        //                 ))
-        //             })?;
-        //
-        //         Ok((content_version, serialised_nrs_map))
-        //     }
-        // };
+        let data = match safe_url.content_version() {
+            None => {
+                self.safe_client
+                    .sequence_get_last_entry(safe_url.xorname(), NRS_MAP_TYPE_TAG, false)
+                    .await
+            }
+            Some(content_version) => {
+                let serialised_nrs_map = self
+                    .safe_client
+                    .sequence_get_entry(
+                        safe_url.xorname(),
+                        NRS_MAP_TYPE_TAG,
+                        content_version,
+                        false,
+                    )
+                    .await
+                    .map_err(|_| {
+                        Error::VersionNotFound(format!(
+                            "Version '{}' is invalid for NRS Map Container found at \"{}\"",
+                            content_version, url,
+                        ))
+                    })?;
+
+                Ok((content_version, serialised_nrs_map))
+            }
+        };
 
         match data {
             Ok((version, nrs_map_xorurl_bytes)) => {
                 // We first parse the NrsMap XOR-URL from the Sequence
-                let url = String::from_utf8(nrs_map_xorurl_bytes.to_owned()).map_err(|err| {
+                let url = String::from_utf8(nrs_map_xorurl_bytes).map_err(|err| {
                     Error::ContentError(format!(
                         "Couldn't parse the NrsMap link stored in the NrsMapContainer: {:?}",
                         err
@@ -343,13 +291,13 @@ impl Safe {
 
                 debug!("Nrs map v{} retrieved: {:?} ", version, &serialised_nrs_map);
                 let nrs_map =
-                    serde_json::from_str(&String::from_utf8_lossy(&serialised_nrs_map.as_slice()))
+                    serde_json::from_str(&String::from_utf8_lossy(serialised_nrs_map.as_slice()))
                         .map_err(|err| {
-                            Error::ContentError(format!(
-                                "Couldn't deserialise the NrsMap stored in the NrsContainer: {:?}",
-                                err
-                            ))
-                        })?;
+                        Error::ContentError(format!(
+                            "Couldn't deserialise the NrsMap stored in the NrsContainer: {:?}",
+                            err
+                        ))
+                    })?;
 
                 Ok((version, nrs_map))
             }
@@ -369,7 +317,7 @@ impl Safe {
     }
 
     // Private helper to serialise an NrsMap and store it in a Public Blob
-    async fn mm_store_nrs_map(&self, nrs_map: &NrsMap) -> Result<String> {
+    async fn store_nrs_map(&self, nrs_map: &NrsMap) -> Result<String> {
         // The NrsMapContainer is a Sequence where each NRS Map version is
         // an entry containing the XOR-URL of the Blob that contains the serialised NrsMap.
         // TODO: use RDF format
@@ -396,7 +344,7 @@ fn validate_nrs_name(name: &str) -> Result<(SafeUrl, String)> {
     }
     // parse the name into a url
     let sanitised_url = sanitised_url(name);
-    let safe_url = Safe::mm_parse_url(&sanitised_url)?;
+    let safe_url = Safe::parse_url(&sanitised_url)?;
     if safe_url.content_version().is_some() {
         return Err(Error::InvalidInput(format!(
             "The NRS name/subname URL cannot contain a version: {}",
@@ -425,19 +373,21 @@ mod tests {
     use anyhow::{anyhow, bail, Result};
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_create() -> Result<()> {
+    async fn test_nrs_map_container_create() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
-        let nrs_xorname = Safe::mm_parse_url(&site_name)?.xorname();
+        let nrs_xorname = Safe::parse_url(&site_name)?.xorname();
 
-        let (xor_url, _, nrs_map) = retry_loop!(safe.mm_nrs_map_container_create(
-            &site_name,
-            "safe://linked-from-site_name?v=0",
-            true,
-            false,
-            false
-        ));
+        let (xor_url, _, nrs_map) = safe
+            .nrs_map_container_create(
+                &site_name,
+                "safe://linked-from-site_name?v=0",
+                true,
+                false,
+                false,
+            )
+            .await?;
 
         assert_eq!(nrs_map.sub_names_map.len(), 0);
         assert_eq!(
@@ -464,7 +414,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_add() -> Result<()> {
+    async fn test_nrs_map_container_add() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
@@ -474,33 +424,19 @@ mod tests {
             .await?;
         let link_v0 = format!("{}?v=0", link);
 
-        let (xorurl, _, nrs_map) = retry_loop!(safe.mm_nrs_map_container_create(
-            &format!("b.{}", site_name),
-            &link_v0,
-            true,
-            false,
-            false
-        ));
-
-        info!("[TESTS] Created mm nrs container at {}", &xorurl);
-
+        let (xorurl, _, nrs_map) = safe
+            .nrs_map_container_create(&format!("b.{}", site_name), &link_v0, true, false, false)
+            .await?;
         assert_eq!(nrs_map.sub_names_map.len(), 1);
         assert_eq!(nrs_map.get_default_link()?, link_v0);
         let _ = retry_loop!(safe.fetch(&xorurl, None));
 
         // add subname and set it as the new default too
         let link_v1 = format!("{}?v=1", link);
-        let (_version, _, _, updated_nrs_map) = retry_loop!(safe.mm_nrs_map_container_add(
-            &format!("a.b.{}", site_name),
-            &link_v1,
-            true,
-            false,
-            false
-        ));
-
-        info!("[TESTS] added to mm nrs container at {}", &xorurl);
-
-        // assert_eq!(version, 1); // Versions features disabled temporarily (TODO replace with hash)
+        let (version, _, _, updated_nrs_map) = safe
+            .nrs_map_container_add(&format!("a.b.{}", site_name), &link_v1, true, false, false)
+            .await?;
+        assert_eq!(version, 1);
         assert_eq!(updated_nrs_map.sub_names_map.len(), 1);
         assert_eq!(updated_nrs_map.get_default_link()?, link_v1);
 
@@ -508,7 +444,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_add_or_remove_with_versioned_target() -> Result<()> {
+    async fn test_nrs_map_container_add_or_remove_with_versioned_target() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
@@ -518,14 +454,15 @@ mod tests {
             .await?;
         let link_v0 = format!("{}?v=0", link);
 
-        let (xorurl, _, _) = retry_loop!(safe
-            .mm_nrs_map_container_create(&format!("b.{}", site_name), &link_v0, true, false, false));
+        let (xorurl, _, _) = safe
+            .nrs_map_container_create(&format!("b.{}", site_name), &link_v0, true, false, false)
+            .await?;
 
         let _ = retry_loop!(safe.fetch(&xorurl, None));
 
         let versioned_sitename = format!("a.b.{}?v=6", site_name);
         match safe
-            .mm_nrs_map_container_add(
+            .nrs_map_container_add(
                 &versioned_sitename,
                 "safe://linked-from-a_b_site_name?v=0",
                 true,
@@ -550,7 +487,7 @@ mod tests {
         };
 
         match safe
-            .mm_nrs_map_container_remove(&versioned_sitename, false)
+            .nrs_map_container_remove(&versioned_sitename, false)
             .await
         {
             Ok(_) => Err(anyhow!(
@@ -574,7 +511,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_remove_one_of_two() -> Result<()> {
+    async fn test_nrs_map_container_remove_one_of_two() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
@@ -584,29 +521,23 @@ mod tests {
             .await?;
         let link_v0 = format!("{}?v=0", link);
 
-        let (xorurl, _, nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_create(
-                &format!("a.b.{}", site_name),
-                &link_v0,
-                true,
-                false,
-                false,
-            )
-        );
+        let (xorurl, _, nrs_map) = safe
+            .nrs_map_container_create(&format!("a.b.{}", site_name), &link_v0, true, false, false)
+            .await?;
         assert_eq!(nrs_map.sub_names_map.len(), 1);
         let _ = retry_loop!(safe.fetch(&xorurl, None));
 
         let link_v1 = format!("{}?v=1", link);
-        let _ = retry_loop!(safe
-            .mm_nrs_map_container_add(&format!("a2.b.{}", site_name), &link_v1, true, false, false)
-        );
+        let _ = safe
+            .nrs_map_container_add(&format!("a2.b.{}", site_name), &link_v1, true, false, false)
+            .await?;
 
-        let _ = retry_loop_for_pattern!(safe.mm_nrs_map_container_get(&xorurl), Ok((version, _)) if *version == 1)?;
+        let _ = retry_loop_for_pattern!(safe.nrs_map_container_get(&xorurl), Ok((version, _)) if *version == 1)?;
 
         // remove subname
-        let (version, _, _, updated_nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_remove(&format!("a.b.{}", site_name), false)
-        );
+        let (version, _, _, updated_nrs_map) = safe
+            .nrs_map_container_remove(&format!("a.b.{}", site_name), false)
+            .await?;
 
         assert_eq!(version, 2);
         assert_eq!(updated_nrs_map.sub_names_map.len(), 1);
@@ -616,7 +547,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_remove_default_soft_link() -> Result<()> {
+    async fn test_nrs_map_container_remove_default_soft_link() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
@@ -626,23 +557,17 @@ mod tests {
             .await?;
         let link_v0 = format!("{}?v=0", link);
 
-        let (xorurl, _, nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_create(
-                &format!("a.b.{}", site_name),
-                &link_v0,
-                true,
-                false,
-                false,
-            )
-        );
+        let (xorurl, _, nrs_map) = safe
+            .nrs_map_container_create(&format!("a.b.{}", site_name), &link_v0, true, false, false)
+            .await?;
         assert_eq!(nrs_map.sub_names_map.len(), 1);
         let _ = retry_loop!(safe.fetch(&xorurl, None));
 
         // remove subname
-        let (_version, _, _, updated_nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_remove(&format!("a.b.{}", site_name), false)
-        );
-        // assert_eq!(version, 1);
+        let (version, _, _, updated_nrs_map) = safe
+            .nrs_map_container_remove(&format!("a.b.{}", site_name), false)
+            .await?;
+        assert_eq!(version, 1);
         assert_eq!(updated_nrs_map.sub_names_map.len(), 0);
         match updated_nrs_map.get_default_link() {
             Ok(link) => Err(anyhow!("Unexpectedly retrieved a default link: {}", link)),
@@ -659,7 +584,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_map_container_remove_default_hard_link() -> Result<()> {
+    async fn test_nrs_map_container_remove_default_hard_link() -> Result<()> {
         let site_name = random_nrs_name();
         let mut safe = new_safe_instance().await?;
 
@@ -669,22 +594,22 @@ mod tests {
             .await?;
         let link_v0 = format!("{}?v=0", link);
 
-        let (xorurl, _, nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_create(
+        let (xorurl, _, nrs_map) = safe
+            .nrs_map_container_create(
                 &format!("a.b.{}", site_name),
                 &link_v0,
                 true,
                 true, // this sets the default to be a hard-link
                 false,
             )
-        );
+            .await?;
         assert_eq!(nrs_map.sub_names_map.len(), 1);
         let _ = retry_loop!(safe.fetch(&xorurl, None));
 
         // remove subname
-        let (version, _, _, updated_nrs_map) = retry_loop!(safe
-            .mm_nrs_map_container_remove(&format!("a.b.{}", site_name), false)
-        );
+        let (version, _, _, updated_nrs_map) = safe
+            .nrs_map_container_remove(&format!("a.b.{}", site_name), false)
+            .await?;
         assert_eq!(version, 1);
         assert_eq!(updated_nrs_map.sub_names_map.len(), 0);
         assert_eq!(updated_nrs_map.get_default_link()?, link_v0);
@@ -692,15 +617,15 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_no_scheme() -> Result<()> {
+    async fn test_nrs_no_scheme() -> Result<()> {
         let site_name = random_nrs_name();
-        let url = Safe::mm_parse_url(&site_name)?;
+        let url = Safe::parse_url(&site_name)?;
         assert_eq!(url.public_name(), site_name);
         Ok(())
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_validate_name() -> Result<()> {
+    async fn test_nrs_validate_name() -> Result<()> {
         let nrs_name = random_nrs_name();
         let (_, nrs_url) = validate_nrs_name(&nrs_name)?;
         assert_eq!(nrs_url, format!("safe://{}", nrs_name));
@@ -708,9 +633,9 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_mm_nrs_validate_name_with_slash() -> Result<()> {
+    async fn test_nrs_validate_name_with_slash() -> Result<()> {
         let nrs_name = "name/with/slash";
-        match validate_nrs_name(&nrs_name) {
+        match validate_nrs_name(nrs_name) {
             Ok(_) => Err(anyhow!(
                 "Unexpectedly validated nrs name with slashes {}",
                 nrs_name
